requirements.txt
pika
kubernetes
argparse
pyyaml
prometheus-client
